+++ rechu/command/new/step/base.py
@@ -13,6 +13,8 @@ from sqlalchemy import inspect
 from sqlalchemy.orm import Session
 from typing_extensions import TypedDict
 
+from rechu.database import Database
+
 from ....io.products import ProductsWriter, SharedFields
 from ....models.product import Product
 from ....models.receipt import ProductItem, Receipt
@@ -64,19 +66,25 @@ class Step(metaclass=ABCMeta):
 
         raise NotImplementedError("Step must be implemented by subclasses")
 
+    @staticmethod
+    def _get_root_product(product: Product) -> Product:
+        return product if product.generic is None else product.generic
+
+    @staticmethod
+    def _is_modified_product(product: Product, session: Session) -> bool:
+        return (
+            cast(int | None, product.id) is None
+            or product in session.dirty
+            or inspect(product).modified
+        )
+
     def _get_products_meta(self, session: Session) -> set[Product]:
         # Retrieve new/updated product metadata associated with receipt items
         return {
-            item.product
-            if item.product.generic is None
-            else item.product.generic
+            self._get_root_product(item.product)
             for item in self.receipt.products
             if item.product is not None
-            and (
-                cast(int | None, item.product.id) is None
-                or item.product in session.dirty
-                or inspect(item.product).modified
-            )
+            and self._is_modified_product(item.product, session)
         }
 
     def _clear_products_meta(self) -> None:
@@ -93,6 +101,7 @@ class Step(metaclass=ABCMeta):
         return {
             product
             for product in unmatched
+            # Check if the product is not empty
             if Product(shop=product.shop).merge(product)
         }
 
@@ -117,7 +126,7 @@ class Step(metaclass=ABCMeta):
             products_writer = ProductsWriter(
                 Path("products.yml"),
                 [
-                    product.generic if product.generic is not None else product
+                    self._get_root_product(product)
                     for product in products
                     if product.generic not in generic_products
                 ],
@@ -141,3 +150,12 @@ class Step(metaclass=ABCMeta):
         """
 
         return False
+
+
+@dataclass
+class DatabaseStep(Step, metaclass=ABCMeta):
+    """
+    A receipt creation step that needs to access the database.
+    """
+
+    database: Database