+++ rechu/command/new/step/meta.py
@@ -15,7 +15,6 @@ from sqlalchemy import select
 from sqlalchemy.sql.functions import min as min_
 from typing_extensions import Required, TypedDict, override
 
-from ....database import Database
 from ....io.products import (
     IDENTIFIER_FIELDS,
     OPTIONAL_FIELDS,
@@ -34,7 +33,7 @@ from ....models.product import (
 )
 from ....models.receipt import ProductItem, Receipt
 from ..input import Input
-from .base import ResultMeta, ReturnToMenu, Step
+from .base import DatabaseStep, ResultMeta, ReturnToMenu
 from .edit import Edit
 from .view import View
 
@@ -74,7 +73,7 @@ MATCHERS: dict[str, _Matcher] = {
 
 
 @dataclass
-class ProductMeta(Step):
+class ProductMeta(DatabaseStep):
     """
     Step to add product metadata that matches one or more products.
     """
@@ -93,7 +92,7 @@ class ProductMeta(Step):
             return {}
 
         # Check if there are any unmatched products on the receipt
-        with Database() as session:
+        with self.database as session:
             self.products.update(self._get_products_meta(session))
             candidates = self.matcher.find_candidates(
                 session, self.receipt.products, self.products
@@ -125,7 +124,7 @@ class ProductMeta(Step):
         return {}
 
     def _load_suggestions(self) -> None:
-        with Database() as session:
+        with self.database as session:
             min_date = session.scalar(select(min_(Receipt.date)))
             if min_date is None:
                 min_date = self.receipt.date
@@ -177,7 +176,7 @@ class ProductMeta(Step):
             product, item=item, initial_key=initial_key, changed=False
         )
         while not matched:
-            changed = initial_product.copy().merge(product)
+            changed = not initial_product.equals(product)
             if initial_key == "!":
                 LOGGER.info("Discarded changes to start with fresh product")
                 _ = product.replace(initial_product)
@@ -203,7 +202,7 @@ class ProductMeta(Step):
                 product, item=item, initial_key=initial_key, changed=changed
             )
 
-        changed = initial_product.copy().merge(product)
+        changed = not initial_product.equals(product)
         if not changed:
             # Should always be existing otherwise it would not match
             LOGGER.info("Product %r remained the same", product)
@@ -213,7 +212,7 @@ class ProductMeta(Step):
         LOGGER.info(
             "Product %s: %r", "updated" if existing else "created", product
         )
-        self.products.add(product)
+        self.products.add(self._get_root_product(product))
         _ = self.matcher.add_map(product)
         if matched_items is not None:
             matched_items.update(matched)
@@ -236,18 +235,21 @@ class ProductMeta(Step):
 
         items = self.receipt.products if item is None else [item]
         matched: set[ProductItem] = set()
-        with Database() as session:
+        with self.database as session:
             self.products.update(self._get_products_meta(session))
             matchers = set(product.range)
             matchers.add(product)
-            if product.generic is not None:
-                matchers.add(product.generic)
+            generic = self._get_root_product(product)
+            matchers.add(generic)
+            matchers.update(generic.range)
 
             pairs = self.matcher.find_candidates(
                 session, items, self.products | matchers
             )
             for meta, match in self.matcher.filter_duplicate_candidates(pairs):
-                if meta in matchers:
+                # Only accept the match if the current product also matches,
+                # but might be overshadowed by generic or sibling range.
+                if meta in matchers and self.matcher.match(product, match):
                     match.product = meta
                     matched.add(match)
                     if not match.discounts and product.discounts:
@@ -256,9 +258,10 @@ class ProductMeta(Step):
                         )
                     else:
                         LOGGER.info("Matched with item: %r", match)
+            self.products.add(self._get_root_product(product))
             self._view_products_meta(
                 "Products that no longer match:",
-                self._update_products_meta(session, self.products | {product}),
+                self._update_products_meta(session, self.products),
             )
 
         return matched, key
@@ -387,7 +390,7 @@ class ProductMeta(Step):
         initial_key = self._set_values(
             product_range, item=item, changed=split_range is not None
         )
-        if initial_key == "" or not initial.merge(product_range):
+        if initial_key == "" or initial.equals(product_range):
             product_range.generic = None
             if split_range is not None:
                 _ = product.merge(split_range)
@@ -404,9 +407,14 @@ class ProductMeta(Step):
         if item is not None:
             LOGGER.info("Receipt product item to match: %r", item)
         else:
-            with Database() as session:
+            with self.database as session:
                 self.products.update(self._get_products_meta(session))
-            _ = View(self.receipt, self.input, products=self.products).run()
+            _ = View(
+                receipt=self.receipt,
+                input=self.input,
+                database=self.database,
+                products=self.products,
+            ).run()
 
         output = self.input.get_output()
         print(file=output)
@@ -441,8 +449,7 @@ class ProductMeta(Step):
         products: tuple[Product, ...] = ()
         if item is None:
             products = tuple(
-                existing if existing.generic is None else existing.generic
-                for existing in self.products
+                self._get_root_product(existing) for existing in self.products
             )
         if product is not None:
             if (product_generic := product.generic) is not None:
@@ -456,7 +463,7 @@ class ProductMeta(Step):
             if item is not None:
                 _ = tmp_file.write(f"# Product to match: {item!r}")
 
-            edit = Edit(self.receipt, self.input, self.matcher)
+            edit = Edit(self.receipt, self.input, self.database, self.matcher)
             edit.execute_editor(tmp_file.name)
 
             reader = ProductsReader(tmp_path)
@@ -483,9 +490,10 @@ class ProductMeta(Step):
     def _edit_extra_products(self, extra_products: tuple[Product, ...]) -> None:
         products = self.products | set(extra_products)
         for extra_product in extra_products:
-            _ = self.matcher.add_map(extra_product)
+            _ = self.matcher.add_map(extra_product, True)
 
-        with Database() as session:
+        with self.database as session:
+            originals = {item: item.product for item in self.receipt.products}
             self._clear_products_meta()
 
             candidates = self.matcher.find_candidates(
@@ -493,13 +501,14 @@ class ProductMeta(Step):
                 self.receipt.products,
                 extra_products,
             )
-            pairs = self.matcher.filter_duplicate_candidates(candidates)
+            pairs = self.matcher.filter_duplicate_candidates(
+                candidates, extra_products
+            )
             for candidate, target in pairs:
-                if (
-                    candidate in products
-                    or candidate.generic in products
-                    or cast(int | None, candidate.id) is None
-                ):
+                new = self._get_root_product(candidate)
+                if new in products or cast(int | None, candidate.id) is None:
+                    if original := originals.get(target):
+                        new.id = self._get_root_product(original).id
                     LOGGER.info("Matching %r to %r", target, candidate)
                     target.product = candidate
 
@@ -515,12 +524,14 @@ class ProductMeta(Step):
             range_index = generic.range.index(product)
             _ = generic.replace(new_product)
             _ = product.replace(generic.range[range_index])
-            generic.range[range_index] = product
-            _ = self.matcher.add_map(generic)
+            product.generic = generic
+            generic.range[range_index:] = [product]
+            _ = self.matcher.add_map(product, True)
+            _ = self.matcher.add_map(generic, True)
         else:
             setattr(new_product, "id", product.id)
             _ = product.replace(new_product)
-            _ = self.matcher.add_map(product)
+            _ = self.matcher.add_map(product, True)
 
     def _get_key(
         self,